The data processing path has been optimised and the interface details have been changed.

+++ Your Grade Book/CustomizedCells/Models/Subject.swift
@@ -6,8 +6,9 @@
 //
 
 import RealmSwift
+
 class Subject: Object {
     @Persisted var name: String
-    @Persisted var maxPoint: Int
+    @Persisted var maxPoint: Double
     @Persisted var tasks: List<Task>
 }
